app/requirements.txt
attrs==23.1.0
blinker==1.6.2
blinker==1.6.2
blinker==1.6.2
blinker==1.6.2
blinker==1.6.2
certifi==2023.5.7
certifi==2023.5.7
charset-normalizer==3.1.0
charset-normalizer==3.1.0
click==8.1.3
click==8.1.3
click==8.1.3
click==8.1.3
click==8.1.3
flasgger==0.9.7.1
Flask==2.3.2
Flask==2.3.2
Flask==2.3.2
Flask==2.3.2
Flask==2.3.2
Flask-JWT-Extended==4.5.2
Flask-JWT-Extended==4.5.2
Flask-JWT-Extended==4.5.2
Flask-Login==0.6.2
Flask-Login==0.6.2
Flask-Login==0.6.2
Flask-Login==0.6.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-SQLAlchemy==3.0.3
Flask-SQLAlchemy==3.0.3
Flask-SQLAlchemy==3.0.3
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
greenlet==2.0.2
greenlet==2.0.2
greenlet==2.0.2
greenlet==2.0.2
idna==3.4
idna==3.4
itsdangerous==2.1.2
itsdangerous==2.1.2
itsdangerous==2.1.2
itsdangerous==2.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
Jinja2==3.1.2
Jinja2==3.1.2
Jinja2==3.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.3
MarkupSafe==2.1.3
MarkupSafe==2.1.3
MarkupSafe==2.1.3
MarkupSafe==2.1.3
mistune==2.0.5
mysql-connector-python==8.0.33
mysql-connector-python==8.0.33
mysql-connector-python==8.0.33
mysql-connector-python==8.0.33
packaging==23.1
protobuf==3.20.3
protobuf==3.20.3
protobuf==3.20.3
protobuf==3.20.3
PyJWT==2.7.0
PyJWT==2.7.0
PyJWT==2.7.0
pyrsistent==0.19.3
PyYAML==6.0
requests==2.31.0
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.15
SQLAlchemy==2.0.15
SQLAlchemy==2.0.15
SQLAlchemy==2.0.15
SQLAlchemy==2.0.15
typing_extensions==4.6.3
typing_extensions==4.6.3
typing_extensions==4.6.3
typing_extensions==4.6.3
typing_extensions==4.6.3
urllib3==2.0.2
urllib3==2.0.2
Werkzeug==2.3.4
Werkzeug==2.3.4
Werkzeug==2.3.4
Werkzeug==2.3.4
Werkzeug==2.3.4
